+++ web/listeMarques.jsp
@@ -242,7 +242,7 @@
                                                         </td>
                                                         <td><%= p.getNom()%></td>
                                                         <td class="btnForm">
-                                                            <a class="btn btn-space btn-secondary" href="./DeleteMarque?id=<%= p.getId()%> ">Supprimer</a>
+                                                            <a onclick="confirmer()" class="btn btn-space btn-secondary" href="./DeleteMarque?id=<%= p.getId()%> ">Supprimer</a>
                                                         </td>
                                                     </tr>
                                                     <% }%>
@@ -257,6 +257,7 @@
                 </div>
             </div>
         </div>
+        <script src="script/confirmDeletion.js" type="text/javascript"></script>
         <script src="assets/vendor/jquery/jquery-3.3.1.min.js"></script>
         <script src="assets/vendor/bootstrap/js/bootstrap.bundle.js"></script>
         <script src="assets/vendor/slimscroll/jquery.slimscroll.js"></script>
